Overhaul of Folder Naming
Decided against IDs as they are too cryptic and unreadable in the
filesystem. I wanted something that could work well even without org.

Now it has a hum readable input for folder names, and a better naming
convention based on the buffer name, and a PROJ-TYPE property that can
be inherited from parent nodes.