Revamped convert to project function
Now this is much more powerful.  It detects the existence of a current
attachment directory and lets the user decide how to move forward with
renaming.  It is able to move the directory on disk in the background
to a new location which will preserve existing links and ensure no
loss of data, while letting the user provide an entirely new project
name.  It does still update the time stamp to reflect when that
project directory name was changed.

A convenience function was also added for simply renaming the
directory.  However, since that is currently the same as converting to
a project, it simply calls this in the background.  The labeling is
only for user convenience so that they know what the function is doing
depending on their use case.